+++ packages/eslint-plugin-kerros/tests/fixtures/bindings.ts
@@ -1,4 +1,0 @@
-import { create } from './reexport'
-import { useSharedModel } from './models'
-
-export const [useShared, SharedProvider] = create(useSharedModel)