+++ packages/eslint-plugin-kerros/tests/setup.ts
@@ -1,8 +1,0 @@
-import { RuleTester } from '@typescript-eslint/rule-tester'
-import { afterAll, describe, it } from 'vitest'
-
-RuleTester.afterAll = afterAll
-RuleTester.describe = describe
-RuleTester.it = it
-RuleTester.itOnly = it.only
-RuleTester.itSkip = it.skip